Add automatic addresses exchange between client and server.
In order to avoid writing the explicit MAC address and IP address of the nodes involved in the communication, an automatic exchange system is implemented.
Just the IP address of the server is required: the client will send a packet with its addresses, and the server will reply with its own addresses.
This way, no hard-coding or configuration is required, as the server IP address is provided as a command-line argument.