Add buffered map between XDP and userspace for kernel bypass.
Transform the packets XDP map from a single-valued map to a "circular queue"-like map. This way, packets are not immediately overwritten but are buffered in the map, allowing the userspace to read them even if the polling is slower than the reception of packets.

It is clear that thanks to memory mapping of the eBPF map, polling from userspace is still much faster than the reception of packets; however, because of any spike in delay between two-polls (which is possible when syscalls are involved), some packet could be lost. Using a buffer, the packet is not overridden, and the userspace app has time to poll the old and the new packets.

The implementation is lock-free: since it is a single-producer single-consumer queue, there should be no need for locks.
Most of the time, XDP and userspace will be handling two different sections of the same memory location, creating no hazard. Although extremely unlikely, in a completely unsynchronized implementation there is the possibility of corruption of packets in the case of XDP writing the same space in memory that userspace is reading.
To address this issue, a "magic number" was added at the end of the pingpong payload, which helps recognize the integrity of the packet without need of any checksum: before reading a packets from the map, the userspace app verified that the magic number matches; after reading a packet, its memory space in the eBPF map is zeroed. By doing this, the only possible way to read a packet in userspace is for XDP to have completely written the packet, including the magic number which is placed at the end; in the case of a half-written packet, the magic number would not match stopping the userspace from reading it.