Add initial AF_XDP program.
The code is adapted from xdp-tools example on AF_XDP.
A lot of the original code was not useful and thus removed.

Currently working but not testable, since XDP forwards only packets with Ethernet type 0x2002, which can only be sent being custom-made by a program, and the current pp_sock does not send packets yet.