Improve abstraction of sending packet thread by adding custom send function.
Different types of pingpong might need different types of send functions. For example, pp_poll sends the packets using a normal socket, while pp_sock sends the packets using AF_XDP socket, which has different functions and dynamics compared to a normal socket.
In order to allow this customization, while maintaining the abstraction over the sending thread, a custom function can be passed to the send thread which will be executed when a packet needs to be sent. Some custom variables can be passed to the send thread in order to be used in the send function (for example, to pass the socket to be used to sent the packet).