Add send and timestamping functionalities to AF_XDP pingpong.
The pingpong using AF_XDP now works, with a customizable amount of packets exchanged at a customizable rate. There is a first phase, just like in pp_poll, in which server and client exchange MAC and IP addresses. Then, the client's send thread starts sending packets at regular intervals using the TX ring of the socket. After the ping is ponged, the timestamps are currently printed to the console.

Currently possible improvements:
- Write the timestamps to file using the persistence agent.
- There is a possible improvement in the packet memory management: the packet headers are written only once as a "template" and then modified for each packet to be sent with the ID; before being sent it's copied to a UMEM frame. There must be a copy: I'm confident it's faster to memcpy the template packet to the UMEM frame than re-creating every time the packet headers on the frame. However, a possible optimization would be instead of writing the pingpong payload on the buffer and the copying the buffer to the UMEM frame, just copy the buffer to the UMEM and then just write the payload information on the UMEM frame. This would avoid the double-write of the payload on both the buffer and the UMEM frame. However, I really think it does not matter, since we are talking about 44 bytes :P.