Fix synchronization issue in UD pingpong.
Since the structure of the original pingpong is very different (in the original PP, packets were sent "synchronously" aka new round starts when previous finishes), some issues have arisen. One being synchronizing access to receive/send buffers: if on the server a packet is being sent back, a new packet is received and copied to the send buffer before the previous packet was sent, there is a data race. To avoid that, before sending a packet there must be no outgoing packet.