Update UD pingpong post 1 recv at a time.
There was a synchronization problem in the recv side: since the send is asynchronous and independent of the state of the system, the recv side would happen to receive two messages too close to each other, overwriting each other's memory, since this basic pinpgong structure only allows for one shared buffer to be used.
To avoid such problem, there is only one RDMA RECV WQE outstanding at a time, having the next one posted when the previous packet is handled. This makes sure that the receive side only has to handle one packet at a time. However, this also implies that packets will be dropped, since the client will send SEND requests that will not be able to be matched with a RECV.
A possible solution to this packet loss is to implement a queueing system similar to XDP.

EDIT: I actually ended up implementing the queueing system but, even if RECVs are available, packets are loss. I have no other option but to blame a packet loss inside the network.