Set AF_XDP to ZeroCopy and fix synchronization issue.
Force the zero-copy mode of AF_XDP. I could not make it work in Linux 5.15.0; in Linux 6.6.19 it works properly.
Introduce a spinlock to avoid synchronization problems in the client, since it accesses the rings both for send and receive from two different threads. Making the send and receive functions critical sections solved the problem.