Implement support for HW timestamps.
When inserting the XDP hook, the program must be flagged as "device-bound" in order to access packet metadata (including HW timestamps).
This can be done by setting: 1. an ifindex that the program is bounded to; 2. the flag `BPF_F_XDP_DEV_BOUND_ONLY`, which I guess tells the kernel that the program is linked to an ifindex not because it's supposed to be offloaded, but because it's just device-bound.