Fix synchronization in queued RDMA UD.
When the server receives a packet from the client, a RECV WR associated with a specific queue index is consumed; the packet is then processed and a send WR is posted to send it back to the client. After that, however, a new recv WR cannot be posted for the same queue idx, because the send operation is still pending: the recv can be posted only when the send operation for that queue index is completed and processed from the completion queue. However, this means that there is the need to keep track of `QUEUE_SIZE` send WR and the send synchronization system becomes slightly more cumbersome.